Solution found
To solve to problem of having to create multiple files, I have decided to create a temporary file that will be used until image view sets the image. Once the image is set, the file is deleted. So instead of having to create n number of files for all the files in the database, only one is created, and is then deleted right after.